requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.0
psycopg2-binary==2.9.10
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
telepot==12.7
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
